move some components to Search, Soorah and Ayah page folders
