+++ src/app/[soorah]/SoorahAyah/SoorahAyah.tsx
@@ -1,37 +1,31 @@
+import type { JSX } from 'react'
+
 import Link from 'next/link'
 import { FaExternalLinkAlt } from 'react-icons/fa'
 
-import Highlighter from 'react-highlight-words'
-
+import { Sajda } from '@/components/Sajda/Sajda'
 import { DisplayData } from '@/helpers/types'
 import { buildUrl } from '@/utility/buildUrl'
 
-import { Sajda } from '../Sajda/Sajda'
-
-export type SearchAyahProps = {
+export type SoorahAyahProps = {
   data: DisplayData
   sajda?: number[]
-  mark?: string
 }
 
-export const SearchAyah = ({ data, sajda, mark = '' }: SearchAyahProps) => (
+export const SoorahAyah = ({ data, sajda }: SoorahAyahProps): JSX.Element => (
   <li className="soorah-list-item">
     <div className="flex flex-row">
       <span className="badge">
-        {data.soorah}:{data.ayah}
+        {data.ayah}
         {sajda?.includes(data.ayah) && <Sajda />}
-      </span>{' '}
-      <Highlighter
-        searchWords={[mark]}
-        textToHighlight={data.content}
-        autoEscape={true}
-        highlightClassName="bg-warning"
-      />
+      </span>
+      <span>{data.content}</span>
     </div>
     <Link
       href={buildUrl(data.soorah, data.ayah, data.translator)}
       className="read-ayah"
       prefetch={false}
+      title="oxu"
     >
       <FaExternalLinkAlt />
     </Link>
