+++ src/app/[soorah]/[ayah]/page.tsx
@@ -3,12 +3,13 @@ import { notFound, redirect } from 'next/navigation'
 import { sirasayi } from 'sirasayi'
 
 import { soorahList } from '@/assets/soorah-list-object'
+import { Bismillah } from '@/components/Bismillah/Bismillah'
 import { ColoredText } from '@/components/ColoredText/ColoredText'
-import { PaginateAyah } from '@/components/PaginateAyah/PaginateAyah'
 import { soorahAyahTitle } from '@/helpers/soorahAyahTitle'
 import { getView } from '@/utility/getView/getView'
 
 import { getAyahService } from './getAyahService'
+import { PaginateAyah } from './PaginateAyah/PaginateAyah'
 
 type AyahProps = {
   params: Promise<{
@@ -64,6 +65,8 @@ const AyahPage = async (props: AyahProps) => {
 
   return (
     <>
+      <Bismillah />
+
       <li className="prose !max-w-none ayah-list-item flex flex-col">
         <span className="text-gray-400">{soorahAyahTitle(soorah, ayah)}</span>
         {content}
