+++ src/app/[soorah]/loading.tsx
@@ -1,5 +1,5 @@
-import { LoadingBoxes } from '@/ui/LoadingBoxes/LoadingBoxes'
+import { LoaderDots } from '@/components/LoaderDots'
 
-export const SoorahLoader = () => <LoadingBoxes />
+export const SoorahLoader = () => <LoaderDots />
 
 export default SoorahLoader
