+++ src/app/search/[search]/SearchAyah/SearchAyah.tsx
@@ -1,32 +1,36 @@
-import type { JSX } from 'react'
-
 import Link from 'next/link'
 import { FaExternalLinkAlt } from 'react-icons/fa'
 
+import Highlighter from 'react-highlight-words'
+
+import { Sajda } from '@/components/Sajda/Sajda'
 import { DisplayData } from '@/helpers/types'
 import { buildUrl } from '@/utility/buildUrl'
 
-import { Sajda } from '../Sajda/Sajda'
-
-export type SoorahAyahProps = {
+export type SearchAyahProps = {
   data: DisplayData
   sajda?: number[]
+  mark?: string
 }
 
-export const SoorahAyah = ({ data, sajda }: SoorahAyahProps): JSX.Element => (
+export const SearchAyah = ({ data, sajda, mark = '' }: SearchAyahProps) => (
   <li className="soorah-list-item">
     <div className="flex flex-row">
       <span className="badge">
-        {data.ayah}
+        {data.soorah}:{data.ayah}
         {sajda?.includes(data.ayah) && <Sajda />}
-      </span>
-      <span>{data.content}</span>
+      </span>{' '}
+      <Highlighter
+        searchWords={[mark]}
+        textToHighlight={data.content}
+        autoEscape={true}
+        highlightClassName="bg-warning"
+      />
     </div>
     <Link
       href={buildUrl(data.soorah, data.ayah, data.translator)}
       className="read-ayah"
       prefetch={false}
-      title="oxu"
     >
       <FaExternalLinkAlt />
     </Link>
