+++ src/app/search/[search]/page.tsx
@@ -8,8 +8,9 @@ import useSWR from 'swr'
 import { ResponseProps } from '@/app/api/v2/search/route'
 import { SOORAH_LIST } from '@/assets/soorah-list-object'
 import { Pagination } from '@/components/Pagination/Pagination'
-import { SearchAyah } from '@/components/SearchAyah/SearchAyah'
 import { fetcher } from '@/utility/fetcher'
+
+import { SearchAyah } from './SearchAyah/SearchAyah'
 
 const Search = () => {
   const params = useParams()
