+++ src/components/LoaderDots.tsx
@@ -1,0 +1,9 @@
+export const LoaderDots = () => (
+  <div className="w-full flex justify-center">
+    <div className="flex space-x-2">
+      <div className="w-2.5 h-2.5 bg-green-600 rounded-full animate-pulse"></div>
+      <div className="w-2.5 h-2.5 bg-green-600 rounded-full animate-pulse delay-200"></div>
+      <div className="w-2.5 h-2.5 bg-green-600 rounded-full animate-pulse delay-400"></div>
+    </div>
+  </div>
+)
